ztp/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pynetbox==7.4.1
python-dotenv==1.0.0
jinja2==3.1.4
ncclient==0.6.15
netmiko==4.4.0
gitpython==3.1.43
aiofiles==24.1.0
python-multipart==0.0.12
